Forecast tests added

+++ Vahetulemus2/AutoTestimine/src/main/java/openWeather/weatherReport/CurrentWeatherReport.java
@@ -3,9 +3,9 @@ package openWeather.weatherReport;
 import openWeather.weather.Coordinates;
 
 public class CurrentWeatherReport extends WeatherReport {
-    public final double tempCurrent;
+    public final long tempCurrent;
 
-    public CurrentWeatherReport(String cityName, Coordinates coordinates, String countryCode, double tempCurrent) {
+    public CurrentWeatherReport(String cityName, Coordinates coordinates, String countryCode, long tempCurrent) {
         super(cityName, coordinates, countryCode);
         this.tempCurrent = tempCurrent;
     }
